Templars are now global range towers added a lot more levels and optimisation fixed the bug when tower would fire at random place instead at the enemy

+++ TowerDefense/Range.swift
@@ -12,7 +12,7 @@ import SpriteKit
 
 protocol Range {
     
-    func getRange() -> CGFloat
+    func getRange() -> Int
     func getNode() -> SKShapeNode
     func getTower() -> Tower
